+++ docs/CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validate impact and summary (OSIDB-1164)
 - Implement tracker description generation (OSIDB-1173)
 
+### Changed
+- Reduce the total amount of records per page when querying Bugzilla (OSIDB-1232)
+
 ## [3.4.1] - 2023-08-21
 ### Changed
 - Fix FlawCollector to account for an empty acknowledgment affiliation (OSIDB-1195)